requirements.txt
flask
psycopg2-binary
paho-mqtt
numpy
pandas
SQLAlchemy
python-dotenv
gunicorn
pytest
